### 2022-01-11 #4047 @whedon
[PRE REVIEW]: visxhclust: An R Shiny package for visual exploration of hierarchical clustering
**Submitting author:** @rhenkin (<a href="http://orcid.org/0000-0002-5511-5230">Rafael Henkin</a>)
**Repository:** <a href="https://github.com/rhenkin/visxhclust" target ="_blank">https://github.com/rhenkin/visxhclust</a>
**Version:** 1.0.0
