### 2022-01-11 #4047 @danielskatz
@rhenkin - thanks for your submission.  I'll look for an editor next.

In addition, you could work on the possibly missing DOI that whedon suggests, but note that this may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references from branch paper` to check again, and the command `@whedon generate pdf from branch paper` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
