### 2022-01-17 #4047 @jonjoncardoso
> @jonjoncardoso @gagolews; would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @osorensen , I'm happy to carry a review of this submission
