### 2022-01-19 #4074 @wiljnich
I'm pleased with this work and have no acceptance-blocking comments. The only outstanding concern I have is the lack of a reference to the documentation page (rhenkin.github.io/visxhclust/) in the README, because several of the key elements of the documentation (e.g. how to open an issue, examples analyses) are exclusively on that site. I would recommend either referring the user to that site via your README, or duplicating several of those elements in the README.

This is great work and I'm happy to see it move towards publication.
