### 2022-01-30 #4074 @jonjoncardoso
This is a great package, the installation and example worked and it is a useful dashboard to explore data visually.

The only thing I felt lacking was a CONTRIBUTING.md so as to adhere to the journal's checklist: _Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support_

Other than that, I am happy to accept this work.
