+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+# Suggestions
+
+If you want to suggest the inclusion or change of a feature, you can either 1) [fork](https://github.com/rhenkin/visxhclust/fork) the repository to work on it and create an [issue](https://github.com/rhenkin/visxhclust/issues/new) to discuss it before proceeding with a pull request, or 2) create an [issue](https://github.com/rhenkin/visxhclust/issues/new) with your suggestion for others to discuss and potentially work on it.
+
+## Assessing suggestions
+
+Major feature changes or suggestions will be considered based on the following criteria:
+
+1. Scope of the tool: do they significantly expand the scope of the tool and/or are out of place relative to the other features?
+1. Visual clutter: do the changes or new inclusions add clutter to the tool or are difficult be visualized with accessible plots?
+1. Computation time: are the changes compatible with a fast response time or a progress-bar approach to compensate waiting time?
+
+Minor changes will be discussed following the same criteria but in a less stringent manner.
+
+# Reporting bugs or asking for help
+
+Please report any bugs or ask for help by creating a new [issue](https://github.com/rhenkin/visxhclust/issues/new).
