+++ README.Rmd
@@ -59,8 +59,6 @@ The app includes multiple help points in the interface (look for the question ma
 - An example of how to [reproduce an analysis](https://rhenkin.github.io/visxhclust/articles/clusterworkflow.html) an analysis using the functions exported by the package. See with `vignette("clusterworkflow")` in R.
 - An example of how to [reproduce the evaluation workflow](https://rhenkin.github.io/visxhclust/articles/clusterevaluation.html) using the functions exported by the package. See with `vignette("clusterevaluation")` in R.
 
-The app includes multiple help points in the interface and a vignette documenting how to load data and the basic clustering loop -- see [vignette("visxhclust")](https://rhenkin.github.io/visxhclust/articles/visxhclust.html).
-
 ## Usage tips and data requirements
 
 To use your data with the tool, you can save a data frame or tibble in an RDS file, or use comma or tab-delimited files, with .csv, .tsv or .txt extensions. The clustering method supported by the tool works only on numeric values; columns containing text will be set aside to annotate the heatmap if so desired. If a column named `ID` exists, it will be used as an internal identifier for rows.
