### 2022-01-31 #4074 @rhenkin
Thanks for the feedback!
@wiljnich I have updated the README.md, having fixed the links to the documentation guides while also changing a bit how it's structured (added a getting started section).
@jonjoncardoso I added a CONTRIBUTING.md file and I also linked to do it in the README.md file.
@osorensen I addressed the main issue and @wiljnich's comment, all changes are in the main repo (not the paper branch).
