### 2022-02-02 #4074 @osorensen
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1093/bioinformatics/btv428 is OK
> - 10.1007/978-3-540-70956-5_7 is OK
> - 10.1111/1467-9868.00293 is OK
> - 10.1093/nar/gkv468 is OK
>
> MISSING DOIs
>
> - 10.1093/bioinformatics/btw313 may be a valid DOI for title: Complex heatmaps reveal patterns and correlations in multidimensional genomic data
>
> INVALID DOIs
>
> - None
> ```

@rhenkin, could you please check if the DOI suggested by whedon is correct, and if so, add it to the paper?
