+++ paper/paper.bib
@@ -73,6 +73,7 @@
   journal = {Journal of the Royal Statistical Society: Series B (Statistical Methodology)}
 }
 @Article{Gu2016,
+    doi = {10.1093/bioinformatics/btw313},
     title = {Complex heatmaps reveal patterns and correlations in multidimensional genomic data},
     author = {Zuguang Gu and Roland Eils and Matthias Schlesner},
     journal = {Bioinformatics},
