### 2022-02-02 #4074 @osorensen
@rhenkin, there is a typo in the first word of the Summary section "visxchlust --> visxhclust"

![image](https://user-images.githubusercontent.com/21175639/152129369-40d5bee1-b352-4010-abe9-384ea4fee5f7.png)
