### 2022-02-02 #4074 @osorensen
@rhenkin, some correction of the reference list:

- Please make sure R is spelled with a capital letter. Adding curly brackets around the word "R" in your `.bib` file should fix it. For example, change the [reference to Shiny](https://github.com/rhenkin/visxhclust/blob/3503e2eb61b1d084122f3b00b4530f5c661ea44c/paper/paper.bib#L56-L62) to
```
@Manual{Shiny,
    title = {shiny: Web Application Framework for {R}},
    author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
    year = {2021},
    note = {R package version 1.7.1},
    url = {https://CRAN.R-project.org/package=shiny},
  }
  ```

This also applies to

> Galili, T. (2015). Dendextend: An r package for visualizing, adjusting, and comparing trees of hierarchical clustering. Bioinformatics. https://doi.org/10.1093/bioinformatics/btv428

and

> Nijs, V. (2021). Radiant: Business analytics using r and shiny. https://cran.r-project.org/ web/packages/radiant/index.html
