+++ paper/paper.bib
@@ -7,7 +7,7 @@
 }
 @Article{Galili2015,
     author = {Tal Galili},
-    title = {dendextend: an R package for visualizing, adjusting, and comparing trees of hierarchical clustering},
+    title = {dendextend: an {R} package for visualizing, adjusting, and comparing trees of hierarchical clustering},
     journal = {Bioinformatics},
     year = {2015},
     doi = {10.1093/bioinformatics/btv428},
@@ -54,7 +54,7 @@
   year = 1936
 }
 @Manual{Shiny,
-    title = {shiny: Web Application Framework for R},
+    title = {shiny: Web Application Framework for {R}},
     author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
     year = {2021},
     note = {R package version 1.7.1},
@@ -87,7 +87,7 @@
     url = {https://CRAN.R-project.org/package=dunn.test},
   }
 @Manual{Nijs2021,
-    title = {radiant: Business Analytics using R and Shiny},
+    title = {radiant: Business Analytics using {R} and Shiny},
     author = {Vincent Nijs},
     year = {2021},
     note = {R package version 1.4.1},
