+++ paper/paper.md
@@ -23,19 +23,19 @@ bibliography: paper.bib
 
 # Summary
 
-**visxchlust** is an R Shiny [@Shiny] web app that facilitates iterative exploration of hierarchical clustering. The package assembles together the outputs of different steps of a clustering workflow in a fluid visual analytics [@Keim] interface, enabling analysts to change selected features and hyperparameters and quickly re-evaluate results. It focuses on hierarchical clustering, a widely used method that is often chosen due to the familiarity of analysts with the tree-like structure of results presented as dendrograms and its frequent combination with visualization techniques such as heatmaps. Package reference and an illustrated tutorial are available [online](https://rhenkin.github.io/visxhclust/articles/visxhclust.html).
+**visxhclust** is an R Shiny [@Shiny] web app that facilitates iterative exploration of hierarchical clustering. The package assembles together the outputs of different steps of a clustering workflow in a fluid visual analytics [@Keim] interface, enabling analysts to change selected features and hyperparameters and quickly re-evaluate results. It focuses on hierarchical clustering, a widely used method that is often chosen due to the familiarity of analysts with the tree-like structure of results presented as dendrograms and its frequent combination with visualization techniques such as heatmaps. Package reference and an illustrated tutorial are available [online](https://rhenkin.github.io/visxhclust/articles/visxhclust.html).
 
 # Statement of need
 
 Clustering methods are supported by dozens of packages in popular data analysis languages such as R and Python. In R, basic clustering functions are included with standard R installations [@Rbase] and are extended by many packages (e.g. `dendextend` [@Galili2015] to modify dendrograms). In Python, state-of-the-art packages such as scikit-learn [@scikit-learn] also include considerable support for clustering. In the R ecosystem, packages such as Radiant [@Nijs2021] and ClustVis [@Metsalu2015] contain Shiny apps that include hierarchical clustering and heatmaps. The first includes a limited set of plots of clustering results and evaluation, while the second has a stronger focus on the pre-processing steps with dimensionality reduction using PCA and does not include evaluation or analysis of the computed clusters. 
 
-The visxchlust package addresses the overheads and difficulties that emerge when analysts have to combine multiple packages to explore different configurations of hyperparameters and features and additionally evaluate and interpret results. It integrates computation, plotting and evaluation of hierarchical clustering in a visual analytics interface by building on state-of-the-art R packages. The interface enables fast iterative workflows without requiring users to learn how to use specific packages or manipulate inputs and outputs of different packages. While the web app does not require any programming knowledge at all to be used, the package also exports internal functions for reproducibility and preparation of figures for publication.
+The visxhclust package addresses the overheads and difficulties that emerge when analysts have to combine multiple packages to explore different configurations of hyperparameters and features and additionally evaluate and interpret results. It integrates computation, plotting and evaluation of hierarchical clustering in a visual analytics interface by building on state-of-the-art R packages. The interface enables fast iterative workflows without requiring users to learn how to use specific packages or manipulate inputs and outputs of different packages. While the web app does not require any programming knowledge at all to be used, the package also exports internal functions for reproducibility and preparation of figures for publication.
 
 # Overview
 
[…]
