### 2022-02-02 #4074 @rhenkin
@osorensen the release version is v1.0.0, the DOI on Zenodo is 10.5281/zenodo.5946818 https://doi.org/10.5281/zenodo.5946817
I fixed the typos and references and generated the PDF again above
