### 2022-02-05 #4074 @Kevin-Mattheus-Moerman
@rhenkin @osorensen I have read the paper and inspected the archive and all seems in order to proceed with acceptance in JOSS.
