### 2022-02-05 #4074 @rhenkin
> Congratulations @rhenkin on getting this work published in JOSS! 🎉
>
> Thanks for editing @osorensen, and thanks @jonjoncardoso and @wiljnich for your review efforts.

Thanks @Kevin-Mattheus-Moerman! Also many thanks to @osorensen @jonjoncardoso @wiljnich
