+++ README.Rmd
@@ -17,6 +17,7 @@ knitr::opts_chunk$set(
 <!-- badges: start -->
 [![R-CMD-check](https://github.com/rhenkin/visxhclust/workflows/R-CMD-check/badge.svg)](https://github.com/rhenkin/visxhclust/actions)
 [![CRAN status](https://www.r-pkg.org/badges/version/visxhclust)](https://CRAN.R-project.org/package=visxhclust)
+[![DOI](https://joss.theoj.org/papers/10.21105/joss.04074/status.svg)](https://doi.org/10.21105/joss.04074)
 <!-- badges: end -->
 
 visxhclust is a package that includes a Shiny application for **vis**ual e**x**ploration of **h**ierarchical **clust**ering. It is aimed at facilitating iterative workflows of hierarchical clustering on numeric data. For that, the app allows users to quickly change parameters and analyse and evaluate results with typical heatmaps with dendrograms and other charts. Additionally, it includes lightweight data overview plots such as correlation heatmaps, annotated MDS and PCA plots. On the evaluation side, it builds on existing packages to compute internal validation scores and Gap statistic, as well as Dunn's test to evaluate significant differences between clusters. Many of the functions are also exported to facilitate documenting a complete analysis cycle.
